chore: 더미 데이터 생성 방식 변경
